requirements.txt
streamlit==1.31.1
python-dotenv==1.0.1
langchain==0.1.14
langchain-google-genai==0.0.9
google-generativeai==0.3.2
google-auth==2.25.2
protobuf==4.25.2
requests==2.31.0
